fix imports, delete duplicated files, health to core
