+++ packages/velo-external-db-core/lib/health/app_info.js
@@ -7,8 +7,6 @@ const maskSensitiveData = (cfg) => {
 
 const appInfoFor = async(operationService, configReaderClient) => {
     const connectionStatus = await operationService.connectionStatus()
-    // const config = await configReaderClient.readConfig()
-    // console.log(configReaderClient);
     const { message: configReaderStatus, authorizationMessage: authorizationConfigStatus } = await configReaderClient.configStatus()
     
     return {
