+++ packages/velo-external-db/test/gen.js
@@ -98,28 +98,6 @@ const systemFieldsWith = fields => {
 }
 
 
-const randomKeyObject = (obj) => {
-    const objectKeys = Object.keys(obj)
-    const selectedKey = objectKeys[Math.floor(Math.random() * objectKeys.length)]
-    return selectedKey
-}
 
-const deleteRandomKeyObject = (obj) => {
-    const deletedKey = randomKeyObject(obj)
-    delete obj[deletedKey]
-    return { deletedKey, newObject: obj }
-}
 
-const randomMatchesValueWithDashes = () => {
-    const num = chance.natural({ min: 2, max: 5 })
-    const arr = []
-    for (let i = 0; i < num; i++) {
-        arr.push(chance.word())
-    }
-    return arr.join('-')
-}
-
-module.exports = {
-    randomDbEntities, randomDbEntity, randomNumberDbEntity, randomNumberColumns, randomColumn, randomCollectionName,
-    systemFieldsWith, deleteRandomKeyObject, randomMatchesValueWithDashes, randomObjectColumn, randomObjectDbEntity
-}
+module.exports = { randomDbEntities, randomDbEntity, randomNumberDbEntity, randomNumberColumns, randomColumn, randomCollectionName, systemFieldsWith }
